Replaced unsigned with size_t to prevent annoying x64/x86 build warnings. Shouldn't really make a difference.

+++ monitor/ActionMonitor/Actions.h
@@ -82,7 +82,7 @@ protected:
   void ClearSearch();
 
   // the command currently chosen
-  unsigned m_uCommand;
+  size_t m_uCommand;
 
 public:
   Action& getCommand( STD_TSTRING* cmdLine = NULL );
